+++ src/main/java/com/pleosoft/feign/acs/search/model/ContentInfo.java
@@ -34,7 +34,7 @@ public class ContentInfo {
 	private String mimeTypeName = null;
 
 	@JsonProperty("sizeInBytes")
-	private Integer sizeInBytes = null;
+	private Long sizeInBytes = null;
 
 	@JsonProperty("encoding")
 	private String encoding = null;
@@ -82,7 +82,7 @@ public class ContentInfo {
 		this.mimeTypeName = mimeTypeName;
 	}
 
-	public ContentInfo sizeInBytes(Integer sizeInBytes) {
+	public ContentInfo sizeInBytes(Long sizeInBytes) {
 		this.sizeInBytes = sizeInBytes;
 		return this;
 	}
@@ -94,11 +94,11 @@ public class ContentInfo {
 	 **/
 	@NotNull
 
-	public Integer getSizeInBytes() {
+	public Long getSizeInBytes() {
 		return sizeInBytes;
 	}
 
-	public void setSizeInBytes(Integer sizeInBytes) {
+	public void setSizeInBytes(Long sizeInBytes) {
 		this.sizeInBytes = sizeInBytes;
 	}
